requirements.txt
pyperclip==1.7.0
PySimpleGUI==4.6.0
python-chess==0.28.3
pyinstaller